requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.7.0
django-environ==0.8.1
djangorestframework==3.13.1
generics==3.4.0
psycopg2==2.9.2
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
uuid==1.30
whitenoise==5.3.0
